requirements.txt
boto3==1.8.6
botocore==1.11.6
Django=>2.1.6
django-sesame==1.4
django-storages==1.7
docutils==0.14
jmespath==0.9.3
myks-gallery==0.7
Pillow==5.2.0
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
whitenoise==4.0
gunicorn==19.7.1
django-split-settings==0.3.0


